# Add Unmarshal perf API to read without allocation