before:
```
goos: linux
goarch: arm64
pkg: github.com/cilium/ebpf/perf
BenchmarkReader
BenchmarkReader-4   	  293677	      3626 ns/op	     409 B/op	       6 allocs/op
```

after:
```
goos: linux
goarch: arm64
pkg: github.com/cilium/ebpf/perf
BenchmarkReader
BenchmarkReader-4              	  311409	      3507 ns/op	     384 B/op	       2 allocs/op
BenchmarkUnmarshalSlice
BenchmarkUnmarshalSlice-4      	  310951	      3667 ns/op	     288 B/op	       1 allocs/op
BenchmarkUnmarshalerSimple
BenchmarkUnmarshalerSimple-4   	  304730	      3669 ns/op	     288 B/op	       1 allocs/op
```
The one alloc is from `prog.Test` and only present during the benchmark.