@ti-mo 2022-05-02
Hey Bryce, thanks for giving this another spin. I'm a bit concerned about the bigger picture here. I see 210 extra lines of code to maintain for a 0.3% execution speed improvement, with total alloc size improvements in the same order of magnitude. Golfing total allocs/op should not be the goal, I think we need something more impactiful than this to consider it worth it.

A few thoughts:
- Using a `sync.Pool` for storing integers seems rather counterproductive. I suspect this undoes the gains gotten by adding custom unmarshalers.
- Eliminating calls to `binary.Read` in favour of implementing `BinaryUnmarshaler`s should result in execution speed gains, but only noticeably on complex types, since those require reflection. `binary.Read` is relatively fast on builtins, so implementing unmarshalers for those might not be worth the LoC. Ideally, this should be benchmarked this in isolation to make sure.
- Could the benchmarks and changes to tests be split off into separate commits?