@brycekahle 2022-05-02
> Golfing total allocs/op should not be the goal

It might be hard to show the gains from this at a library level. If you want, I could incorporate this change into our agent and put it through our load testing environment.

> with total alloc size improvements in the same order of magnitude.

`BenchmarkReader` is maintaining the previous behavior, which still does a copy and alloc in this library. Using `Unmarshal` results in 0 allocs/op from the library itself, which is a gain of 96 bytes/op and 5 allocs/op.

The main benefit is that the user of the library has total control of the allocation behavior. 

> Using a sync.Pool for storing integers seems rather counterproductive. I suspect this undoes the gains gotten by adding custom unmarshalers.

I'm using the `sync.Pool` for the slice that is necessary to read the `uint32`. This is because it can span the end of the circular buffer, so there is not an easy stack-based way to read the value (my other PR did this, but didn't account for the end of the buffer).

> Eliminating calls to binary.Read in favour of implementing BinaryUnmarshalers should result in execution speed gains

The goal here was reducing allocs, not necessarily speed gains.

> Could the benchmarks and changes to tests be split off into separate commits?

Sure!