@brycekahle 2022-05-03
> If you want, I could incorporate this change into our agent and put it through our load testing environment.

I went ahead and did this. By giving us control of allocations, we are able to save >15% of our RSS/working set memory.