@lmb 2022-05-03
I agree with Timo that the change is pretty intrusive for what is superficially a small gain. Seems like my idea wasn't particularly good, sorry. I have some questions to help me understand what's going on:

* What does your load testing look like? Is it "max throughput" with given resources, or does it simulate a prod deployment?
* Is there a way for us to reproduce your load testing?
* At which point did you take RSS, some steady state? Did you average it or something similar?
* Are you seeing CPU usage go down as well? Any changes to GC pause metrics, or how frequently the GC runs?
* Do you tweak GOGC or the like?