@brycekahle 2022-05-03
Because we are processing data on a continuous basis, excess allocations on the hot path cause the heap size to increase between GC runs. This causes the GC to increase the NextGC value, and ultimately lead to a higher steady state RSS of our service. Reducing allocations has a significant impact, and we have already seen improvements from other strategies like string interning, using the new stack-based IP address type `netaddr.IP`, using `sync.Pool`s, and others.

The allocations from the perf `Read` are not the biggest slice of the pie, but right now we have no ability to control the allocation behavior (without forking, something we'd like to avoid, if possible).

> * What does your load testing look like? Is it "max throughput" with given resources, or does it simulate a prod deployment?

It is a continuous environment where we are sending 3000 HTTP requests/sec at a host. That host is running multiple versions of our ebpf-based agent, so we can compare apples to apples with regard to the network traffic that they see.

> * Is there a way for us to reproduce your load testing?

It would be difficult.

> * At which point did you take RSS, some steady state? Did you average it or something similar?

We measure it every 20s, and plot it. It is not a flat line, but seems to oscillate around a value. We run the experiment for a long time (> 1 hour) to make sure spurious behaviors do not influence the analysis.

> * Are you seeing CPU usage go down as well?

No significant changes to CPU for us.

> Any changes to GC pause metrics, or how frequently the GC runs?

I can take a look, but GC pauses generally aren't the problem for us. This isn't a web server where a GC pause will cause the 99th percentile latency to spike. 

> * Do you tweak GOGC or the like?

We haven't yet, but this is something we will likely do very soon.