@lmb 2022-05-04
I took another look, I think my idea to use an "unmarshal into interface{}" API was bad. Sorry for sending you on a wild goose chase. Another thing I noticed is that the benchmark is dominated by submitting samples into the perf ring. This explains why the gains look small, when in reality they are really quite sizeable for the work we actually care about.

Can you please give https://github.com/lmb/ebpf/tree/reader-read-buffer a try in your test environment and see if it has a similar effect? It's got the same basic optimizations you did, except that I avoided `sync.Pool` in favor of a small preallocated slice in `Reader`. It also has a new method on Reader:

```
func (pr *Reader) ReadBuffer(buf []byte) (Record, error)
    ReadBuffer is like Read except that it allows reusing buffers.

    buf is used as Record.RawSample if it is large enough to hold the sample
    data. If buf is too small a new buffer will be allocated instead. It is
    valid to pass nil, in which case ReadBuffer behaves like Read.
```

There is an example that shows how to use it, if its not clear. The benefit of this API is that we can return `Record, error` instead of `cpu int, lost int, error` like for `Unmarshal`. We also don't have to get into the business of unmarshaling things in the first place.