@brycekahle 2022-05-05
> Can you please give https://github.com/lmb/ebpf/tree/reader-read-buffer a try in your test environment and see if it has a similar effect?

I'm testing this out today and will report back.