@brycekahle 2022-05-05
Tested out your branch, we get similar gains from it. If you are more comfortable taking your approach, then let's go for it.