@lmb 2022-05-06
Thanks for being patient and explaining the problem to me. `ReadBuffer` is the way to go if it satisfies your needs, since its less code overall. Please take a look at #663.